+++ divio_cli/test/test_commands.py
@@ -2,7 +2,18 @@ import shlex
 import subprocess
 
 import pytest
+import os
+import shutil
+from tempfile import TemporaryDirectory
 
+from click.testing import CliRunner
+
+from divio_cli import cli
+from divio_cli.cloud import WritableNetRC
+
+
+NOI = "--noinput"
+DBFILE = ".divio/local_db.sql"
 
 TEST_COMMANDS_CLICK = [
     "doctor",
@@ -38,3 +49,42 @@ def test_call_click_commands(divio_project, command):
     exitcode = subprocess.check_call(["divio", *shlex.split(command)])
 
     assert exitcode == 0
+
+
+@pytest.mark.integration()
+def test_logout_command(divio_project):
+    """
+    The logout command has to be tested separately, and using a temporary
+    netrc, because the following tests require the cli to be logged in.
+
+    This test copies the global netrc into a temporary directory, sets the
+    temporary netrc globally, and then checks if the "logout" command causes
+    the "login --check" command to return an exit code of 1.
+    """
+
+    with TemporaryDirectory() as temporary_directory:
+        original_netrc_path = WritableNetRC.get_netrc_path()
+
+        temporary_netrc_path = os.path.join(
+            temporary_directory,
+            os.path.basename(original_netrc_path),
+        )
+
+        shutil.copy(original_netrc_path, temporary_netrc_path)
+
+        try:
+            os.environ["NETRC_PATH"] = temporary_netrc_path
+
+            runner = CliRunner()
+
+            result = runner.invoke(cli.cli, ["login", "--check"])
+            assert result.exit_code == 0, result.stdout
+
+            result = runner.invoke(cli.cli, ["logout", "--non-interactive"])
+            assert result.exit_code == 0, result.stdout
+
+            result = runner.invoke(cli.cli, ["login", "--check"])
+            assert result.exit_code == 1, result.stdout
+
+        finally:
+            os.environ.pop("NETRC_PATH")