+++ divio_cli/test/test_localdev_utils.py
@@ -1,7 +1,12 @@
 import pytest
 
 from divio_cli import settings
-from divio_cli.exceptions import ConfigurationNotFound
+from divio_cli.exceptions import (
+    ConfigurationNotFound,
+    DivioException,
+    DivioWarning,
+    DockerComposeDoesNotExist,
+)
 from divio_cli.localdev import utils
 
 
@@ -19,3 +24,25 @@ def test_get_application_home(tmp_path):
     p.write_text("#Examplecontent")
     home = utils.get_application_home(str(tmp_path))
     assert home
+
+
+@pytest.mark.parametrize(
+    ("exception", "expected"),
+    [
+        (ValueError(), ""),
+        (RuntimeError("oops"), "oops"),
+        (DivioException(), " error!\n"),
+        (DivioWarning("oops"), " warning!\noops"),
+        (
+            DockerComposeDoesNotExist(),
+            " error!\ndocker-compose.yml does not exist.",
+        ),
+    ],
+)
+def test_timed_step_exception(exception, expected):
+    with pytest.raises(Exception) as excinfo:
+        with utils.TimedStep("test"):
+            raise exception
+
+    assert isinstance(excinfo.value, exception.__class__)
+    assert str(excinfo.value) == expected