+++ CHANGELOG.rst
@@ -8,9 +8,12 @@ x.x.x (x-x-x)
 * Added support to list and add services to an application.
 * Added support to list regions.
 * Added support to list organisations.
-* Improve exception handling.
+* Added support for environment prefix to app pull.
+* Improved exception handling.
 * Removed support for Python 3.6 (end of life since December 2021).
 * Added support for Python 3.10 and Python 3.11.
+* Added the option to keep temporary files for app push and pull.
+* Added logout command.
 
 3.17.2 (2023-03-16)
 -------------------